requirements.txt
matplotlib
hydra-core
wandb
transformers
datasets
seaborn
scipy
timm
detectors
wandb
lightning